removed properties, added some formatting
I took out "properties: {}" added a link and some emoji formatting to inquire about inclusion in other areas.   

I'm into "super learning" techniques - 
I'd suggest that because Bicep is "a zebra with different patterned stripes" (it's beautiful, but unfamiliar) it will always be "one more thing" that most advanced Azure users will need to learn and maintain some level of proficiency with.   I understand the use case for the Zebra -- Bicep + VS Code intelligent is intuitive & friendly, but still requires a time investment to gain proficiency, especially if the user isn't coming equipped with an ARM or Azure Cloud background (like an AWS person might be).   So this is my proposal to lower that barrier to entry. 

I believe that as documentation goes, the inclusion of pictographic symbols (such as Chinese or Emoji), offer clues to the reader helping them identify the nature of the comment they are about to read i.e. "simulating the optical synapses" 
is it a warning ⚠️, a hint 🤓, or a link to a more comprehensive 📖 documentation resource? 

Emoji has a way of making the content easier to visually consume.   
I think the color breaks up the walls of text by providing visual anchors make it easier to digest.  (at least in my brain)

These days I'd put emojis in my resourceGroup names if Azure would let me!  
So feel free to reject or suggest that this is something I should float in a separate Issue?