Known top-level properties does not block compilation if not defined in resource schema
Using properties such as sku, identity, kind on a resource will be treated as warning instead of error if it's not defined in resource type swagger specs.
Additionally, if a user uses a property that is not defined, a warning message has been extended with a link to place collecting user reports on type inaccuracies.