Fix parameter passing for resource->object params
Fixes: #6038

This change fixes a regression introduced by adding resource-typed
parameter support. Resource-typed parameters require an implicit
conversion to pass the resource id as a string rather than the resource
body as an object.

The codegen for parameters was applying the conversion for all cases
where the RHS was a resource without checking for the LHS to be a
resource-typed parameter. The fix is to correct that check.