Support `@metadata` decorator on outputs
The runtime supports having `metadata` on the output objects (and the
`@description` decorator works by adding a value to this) but Bicep
did not allow using the `@metadata` decorator on outputs to add
custom metadata.

This change fixes that issue.

With this change, you can now write:

```bicep
@metadata({
  works: true
})
@description('an example output, with metadata')
output example string = ''
```

Which is complied down to:

```json
"outputs": {
  "example": {
    "type": "string",
    "value": "",
    "metadata": {
      "description": "an example output, with metadata",
      "works": true
    }
  }
}
```

Fixes #6250