Add string literal checking for all-string unions. Add tests.
Remove the check for integer literal. I think string based property access is the only thing that needs to be checked here. If a resource is resolvable and something else is used, it's going to be a type error, so having the dtc diagnostic and type diagnostic visible at the same time will make it easier for the programmer to figure out what to do next.