Add support for defining and emitting user defined discriminated object union types. (#11222)
Fixes #9230 

## Contributing a feature

* [x] I have opened a new issue for the proposal, or commented on an
existing one, and ensured that the Bicep maintainers are good with the
design of the feature being implemented
* [x] I have included "Fixes #{issue_number}" in the PR description, so
GitHub can link to the issue and close it when the PR is merged
* [x] I have appropriate test coverage of my new feature

## Summary
- Adds support for defining discriminated object unions with an explicit
discriminator property set via a type decorator `@discriminator`.
- Adds support for ARM template emission of discriminated object unions.
- Does not support the additional properties feature in discriminated
object unions, including in individual union members.
- Uses existing support for discriminated object union completions (does
not extend it). This is currently limited to just the discriminated
property. Future PR will need to address this.

## Test files

[9230-test.zip](https://github.com/Azure/bicep/files/12071367/9230-test.zip)