Fix nested type property decorator completions (#11972)
Resolves #11970 

The LS is currently looking at the type of the top-level statement in
which a cursor is located when proposing decorator completions. This
means that for a scenario like the following:

```bicep
type obj = {
  @|
  prop: int
}
```

the LS will propose completions for decorators that can be used on
objects (the type of the full `type obj = {...}` statement) rather than
decorators that can be used on integers (the type of the decorator's
target). Consequently, the LS is proposing completions that will raise
an error diagnostic (like `@sealed()` or `@discriminator()`) and not
proposing completions that would be valid (like `@minValue()`).
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/11972)