fix for incorrectly adding back 'Needs: Triage' unnecessarily (#12342)
Fixes #12341 

I'm not sure what the purpose is of the section I removed. The
description is prefixed with `[Closed Issues]` - does that mean it was
only supposed to run on closed issues?

In practice, what the rule does is add back `Needs: Triage` to any issue
with the `Needs: Author feedback` label, but if an issue has that label,
then it doesn't need to be triaged. It needs to sit and wait for the
author to post a response.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/12342)

Co-authored-by: Anthony Martin <38542602+anthony-c-martin@users.noreply.github.com>