#13619 secure parameter linter fix (#13717)
Fixes  #13619

What was happening was that the linter rule was falling into the default
behaviour and using the default error string as it wasn't checking it
was a variable assignment. Added 1 unit test to cover the basic premise
of the lint rule failure.

One thing which did occur to me was that assignment of the secure
parameter to a variable which is a hardcoded string isnt caught - but
seemed like it should be a lint rule in of itself? Thoughts?

eg.

```
var blah string = 'something'

@secure()
param myparam string = blah
```

Should we overcomplicate the lint rule to find a reassignment of a
secure parameter? Seems an illogical but valid thing to do though