Fix Intellisense and other features of Bicep in Visual Studio that were broken by VS 17.10 (#14532)
Fixes https://github.com/Azure/bicep/issues/14316

VS 17.10 added support for language server capabilities. This means that
it started firing notifications only for events that were explicitly
requested by the client (which is normal and works in vscode).
Unfortunately, their LSP implementation lacks support for document
selectors based on language ID (they only support file patterns,
language/scheme are NYI). This means it didn't recognize any of our
registrations and VS was giving us virtually no notifications, e.g. for
file changes. The fix is to use file patterns for VS, language/scheme
for any other hosts including vscode.

I fixed this such that it should only affect the behavior in VS, not
VsCode or other hosts, out of an abundance of caution (it would be
possible to return language/scheme as well as file patterns, but it does
cause minor behavior changes in vscode and would also mean VS behavior
would change in minor ways again after this fix this problem - better to
just make the behavior explicit for both hosts).

The *actual* fix is in DocumentSelectorFactory, almost all of the rest
is just passing along the dependencies it needs to know when to apply
the fix.