Fix go to def for template spec modules (#14871)
The "Go to Definition" feature for template specs in VS Code is
currently broken. When used, it opens the template spec JSON file in an
editable state, though it should be read-only. This action then triggers
the `documentDidOpen` event for the JSON file, which leads to a
compilation update for the template spec module. As a result, the Bicep
compiler mistakenly interprets the template spec JSON file as an ARM
template, causing an `The reference ARM template has errors` message.
This PR aims to fix that issue.

The problem was initially identified during the investigation of issue
#14660. However, this PR might not fully resolve the issue, as according
to the customer, the `The reference ARM template has errors` message can
still appear randomly, even when "Go to Definition" is not used.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/14871)