Expose `fail` function (#15958)
Resolves #14067

This PR exposes the new `fail` function that is available in the ARM
engine. One thing I went back and forth on is what the return type of
this function should be. Three viable options are:
- `Error`
- `Any`
- `Never`

My first instinct was to use `Error`, but this makes the `fail` function
surprisingly difficult to use. It's essentially forbidden everywhere
unless you wrap it in `any()`, which hurts usability. I considered using
a return type of `any` to remove the need for this boilerplate -- this
is close to the semantic used in [the original C# proposal for `throw`
expressions](https://github.com/dotnet/csharplang/blob/main/proposals/csharp-7.0/throw-expression.md)
-- but I wasn't happy with the downstream effects of given Bicep's union
type semantics. If `fail(...)` is of type `any`, then `condition ? 'boo'
: fail('boo!')` is of type `any`, since the union of `any` with anything
else is `any`. This isn't really correct, since the expression will only
be evaluated if the condition is true, in which case the ternary's type
would be `'boo'`.

I ended up choosing `never` both because a `fail(...)` expression can't
ever finish evaluating (and therefore doesn't really have a type) and
because it had the right downstream assignability behavior. Very open to
discussion on this if anyone feels strongly that `never` is incorrect,
though!
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/15958)