Fix 13606, stuck in ext-src loop (#16381)
Fixes #13606

The URI we use to show external bicep sources is encoded to contain the
information necessary as query params. vscode takes the portion before
the query params as a file path and shows the filename in the title, so
we're using that to control how it shows up in the title (not aware of a
better way to do that). Right now, it doesn't end with .bicep so we have
some extension code to force it to bicep language id. That's apparently
flaky, best fix is to modify our title so that the actual filename is at
the end of the path (e.g. main.bicep), then that extension code is no
longer necessary at all.

Example current tab title:

![image](https://github.com/user-attachments/assets/9d740dfe-bda4-4d2d-b0cd-1fcd28bb762e)

Changed examples:

![image](https://github.com/user-attachments/assets/f5f6e639-b164-41c6-94ee-28c95d757bfb)

![image](https://github.com/user-attachments/assets/92eb4051-7b42-4320-a650-29091a57879a)

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/16381)