Forward base-type errors when accessing type properties (#16394)
Resolves #16452 

Type property access is only allowed on resource-derived types and types
declared in `type` statements, but the check enforcing this constraint
will mask any error in the type of the property access's base
expression. This PR makes sure that the constraint is only enforced if
the base expression is of a non-error type so that the most specific and
actionable error message is not swallowed.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/16394)