[Wait-And-Retry] Decorator output standardization (#16603)
This is an incremental PR that introduces the WaitUntil and RetryOn
decorators to the Bicep language.
The PR introduces changes on WaitUntil and RetryOn decorators where the
output of the decorators is standardized based on this PR
https://github.com/Azure/bicep-reps/pull/13


Context:
Users are facing issues with certain RPs (example SRP), where due to
latency in replication or RP itself responded with success even before
the resource is ready for use.

In such a case, dependent resource deployments fail, because the
resource isn't available according to ARM or some properties of the
resource isn't available like list keys for example.
https://github.com/Azure/bicep/issues/1013 for reference.

Previous PRs
https://github.com/Azure/bicep/pull/16167
https://github.com/Azure/bicep/pull/16342
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/16603)

---------

Co-authored-by: Subha Sambandamurthy <susamban@microsoft.com>