Put new DSC scope behind feature flag
And with this, all the unit tests pass because the feature flag is checked in the right places.
Meaning we get BCP033 for the incorrect assignment of `targetScope = 'dsc'`
and BCP135 for the invalid scope on the resource when the feature is not enabled.

Now we need new tests, documentation, and more usage to figure out the path forward.