Use languageVersion 2.0 when an existing resource has an explicit dependency (#16670)
Resolves #16664

In a symbolic name template, an `existing` resource can depend on other
resources or modules, and this will cause the backend to delay reading
the resource until those dependencies are completed. This allows you to
deploy a resource in a module, then refer to it with an `existing`
declaration in the module parent. In non-symbolic name templates,
`existing` resources are compiled to `reference()` expressions and are
all handled in the first wave of deployment jobs.

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/16670)