Add Desired State Configuration target scope and experimental feature flag (#16389)
This PR is a minimal implementation that allows Bicep to transform a
file of DSC resources (denoted with `targetScope = 'dsc'`) and emit an
ARM file compatible with the [DSC]( https://github.com/PowerShell/DSC)
CLI. All functionality (of which there is beautifully minimal) is
guarded by a new experimental feature flag, and can be enabled with a
`bicepconfig.json` like this:

```json
{
  "experimentalFeaturesEnabled": {
    "resourceDSC": true
  }
}
```

Because Bicep is great, when enabled, the VS Code extension properly
understands the new target scope (shows it in completions, doesn't mark
it as the error that it is when not enabled).

## Contributing to documentation

* [x] All documentation contributions should be made directly in the
[Bicep documentation on Microsoft
Docs](https://docs.microsoft.com/azure/azure-resource-manager/bicep/).
-- Nothing to document yet beyond the new flag, which was added to the
relevant document in this repo.

## Contributing an example

We are integrating the Bicep examples into the [Azure QuickStart
Templates](https://github.com/Azure/azure-quickstart-templates/blob/master/1-CONTRIBUTION-GUIDE/README.md).
If you'd like to contribute new example `.bicep` files that showcase
abilities of the language, please follow [these
instructions](https://github.com/Azure/azure-quickstart-templates/blob/master/1-CONTRIBUTION-GUIDE/README.md)
to add them directly there. We can still take bug reports and fixes for
the existing examples for the time being.

* [ ] This is a bug fix for an existing example -- it is not.
* [ ] I have resolved all warnings and errors shown by the Bicep VS Code
extension -- making all of this work well in the extension is the next
part, in a separate PR.
* [x] I have checked that all tests are passing by running `dotnet test`
* [x] I have consistent casing for all of my identifiers and am using
camelCasing unless I have a justification to use another casing style --
I chose to keep 'DSC' capitalized because it is an acronym, so
'resourceDsc' looks silly.

## Contributing a feature

* [x] I have opened a new issue for the proposal, or commented on an
existing one, and ensured that the Bicep maintainers are good with the
design of the feature being implemented -- yes via internal meeting and
whitepaper.
* [ ] I have included "Fixes #{issue_number}" in the PR description, so
GitHub can link to the issue and close it when the PR is merged -- no
open issue for this.
* [ ] I have appropriate test coverage of my new feature -- open
question for maintainers: how much test coverage is appropriate for a
new under-development experimental feature?

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/16389)