Local Deploy: Support deploying Azure modules (#16752)
This PR adds the ability to annotate a .bicep file with `targetScope =
'local'` if the `localDeploy` feature is enabled. This now allows a
mixture of "local" and "azure" Bicep files in a module hierarchy, and
support has been added for deploying "azure" modules as part of a local
deployment.

This introduces a **breaking change** to local-deploy, because it means
that any `.bicep` file using the functionality will need to use the
following:
```bicep
targetScope = 'local'
```
 
The reason for this is that currently we use the enablement of the
experimental feature as a means of determining whether or not to support
"local mode". This will not work if we want to make this functionality
non-experimental - as it's the only way we can disambiguate "local" vs
"azure".

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/16752)