Add linter rule to warn on unused compile-time imports (#16990)
Fixes #11877

When using an alias (`import * as alias from 'mod.bicep'`, a warning
will only be raised if the alias is unused. In the case where functions
or variables that can be accessed through this alias are not used, the
rule is unable to detect those.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/16990)