Bicep Extension Framework Updates
Updated the framework to simplify the extension writing experience. In
this update we introduce the ability to leverage .net dependency
injection for easier extensibility and flexibility.  An effort has been
made to further abstract the integration of gRPC and handlers to make
the experience as simple as possible.  We also introduced a standard
type spec generator that can be used by anyone to generate a basic type.
If the generic spec generator doesn't meet your needs simply implement
the ITypeSepcGenerator interface to tailor the experience to your needs.
The debug experience has been greatly simplified by allowing you to
execute send requests to the gRPC server when running in http mode.
These requests can be sent through tools such as grpculr.

Finally new cli parameters have been added:

--http <port> : http port to use for grpc server; defaults 5000
--describe <true|false>: output the type spec for the extension without
launching the server