Surface cross-module validation errors in snapshot command as messages, not unhandled exceptions (#17378)
Resolves #17322 

Snapshotting can fail when a validation error only arises based on the
parameters passed to a module. This currently surfaces as a messy
unhandled exception with a stack trace, even though it represents an
error that the author will need to correct. This PR keeps the error
message from any `TemplateValidationException`s encountered but ensures
the stack trace is not dumped to stderr.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/17378)