Mark conditional resources as nullable for the purposes of property access (#16987)
Resolves #3750 

This PR updates `.?` is compiled when used on top-level properties of
resources and modules. The compiler will use the condition of the
resource or module as the guard clause in an `if` expression,
substituting `null()` if the condition is not met. For example, in the
following template:

```bicep
param condition bool

resource r 'RP.Namespace/foos@1999-12-31' = if (condition) {
  ...
}

output prop string = r.?properties.bar
```

The vaule of the `prop` output will compile to an expression like
`[tryGet(if(parameters('condition'), reference('r'), null()), 'bar')]`.

Type analysis to make sure a warning is emitted when a `.?` should be
used but is not and to make sure chained accesses are properly typed was
complected by the fact that references are only _situationally_
nullable. For example, given the statement:

```bicep
resource accounts 'Microsoft.Storage/storageAccounts@2023-05-01' = [for i in range(0, 10): if (i % 2 == 0) {
  name: 'sa${i}'
}]
```

then the expression `accounts[1].properties` (which compiles to
`[reference('accounts[1]').properties]`) will cause a deployment
failure, whereas `map(accounts, a => a.properties)` (which compiles to
`[map(references('accounts'), lambda('a',
lambdaVariables('a').properties))]`) will not because the output of
`references()` only includes resources with a `true` or absent
condition. The `dependsOn` property in ARM is also specifically allowed
to refer to resources with a `false` condition, but it _may_ be a
problem if the `parent` or `scope` property in Bicep refers to a
resource with a `false` condition.

I opted to make property access on a base that is a resource nullable.
E.g., `r` in the first example above is not nullable, but it becomes
nullable when used in the expression `r.properties`. Similarly for
collections, `accounts` in the second example is an array of nonnullable
value, but `accounts[<index>]` becomes nullable when used in an
expression like `accounts[<index>].properties`. This seems error prone,
however, so it's probably best to tag this for a team discussion.

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/16987)