File I/O abstraction part 18: Migrate all commands (#17576)
Updated all commands to use `Bicep.IO`.

The second commit contains a breaking change to disallow `\` to be used
in file input and file pattern CLI args on Unix platforms. The reason is
that `\` is never supported in module paths, so allowing the entry point
path to contain `\` is misleading. Besides, `\` is never correctly
handled because we convert the path to `Uri` with `UriBuilder`, but it
normalizes `\` to `/` on all platforms,. On Linux and macOS, the `Uri`
will point to a wrong file.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/17576)