Fix for incorrect error diagnostic on policyDefinitions existing resource usage (#18169)
Quick fixes for issues: [18163
](https://github.com/Azure/bicep/issues/18166)&
[18163](https://github.com/Azure/bicep/issues/18163)

ValidParentScopes must contain scopes that can be written to and read
from. The broken implementation only passed in scopes that could be
written to, hence breaking customers who tried to use the `existing`
keyword with a scope that could be written to and read from.

Root cause of the issue is that
[ResourceTypeComponents](https://github.com/Azure/bicep/blob/2b1a61d6d6dbba88fb6f9dd798be7d2af5cfd90a/src/Bicep.Core/TypeSystem/Types/ResourceType.cs#L11)
has a property named "ValidParentScopes", which is intended to capture
both writable and readable scopes, but the type loading code was updated
to be more precise about differentiating readable + writable scopes, and
misinterpreted this field to just mean "writable scopes" (see
[here](https://github.com/Azure/bicep/blob/2b1a61d6d6dbba88fb6f9dd798be7d2af5cfd90a/src/Bicep.Core/TypeSystem/Providers/Extensibility/ExtensionResourceTypeFactory.cs#L42))
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/Azure/bicep/pull/18169)